moved tabwindow to ui files
