+++ src/ui/CMakeLists.txt
@@ -1,7 +1,6 @@
 add_library(ui STATIC
     ui.cpp ui.h
     scenario_controls.cpp scenario_controls.h
-    tabwindow.cpp tabwindow.h
     bucket_tab.cpp bucket_tab.h
 )
